Refactor out new epoch creation
